REFACTOR Common functions to different file in deviceService
