+++ lib/services/devices/deviceService.js
@@ -30,124 +30,12 @@ var request = require('request'),
     logger = require('logops'),
     config = require('../../commonConfig'),
     ngsiParser = require('./../ngsi/ngsiParser'),
+    registrationUtils = require('./registrationUtils'),
     subscriptions = require('../ngsi/subscriptionService'),
     _ = require('underscore'),
     context = {
         op: 'IoTAgentNGSI.DeviceService'
     };
-
-
-/**
- * Generates a handler for the registration requests that checks all the possible errors derived from the registration.
- * The parameter information is needed in order to fulfill error information.
- *
- * @param {Boolen} unregister       Indicates whether this registration is an unregistration or register.
- * @param {Object} deviceData       Object containing all the deviceData needed to send the registration.
- * @return {Function}              The generated handler.
- */
-function createRegistrationHandler(unregister, deviceData, callback) {
-    return function handleRegistrationResponse(error, response, body) {
-        if (error) {
-            logger.error(context, 'Connection error sending registrations to the Context Broker: %s', error);
-            callback(error);
-        } else if (response && body && response.statusCode === 200) {
-            var errorField = ngsiParser.getErrorField(body);
-
-            if (errorField) {
-                logger.error(context, 'Registration error connecting to the Context Broker: %j', errorField);
-                callback(new errors.BadRequest(JSON.stringify(errorField)));
-            } else {
-                logger.debug(context, 'Registration success.');
-                callback(null, body);
-            }
-        } else {
-            var errorObj;
-
-            logger.error(context, 'Protocol error connecting to the Context Broker: %j', errorObj);
-
-            if (unregister) {
-                errorObj = new errors.UnregistrationError(deviceData.id, deviceData.type);
-            } else {
-                errorObj = new errors.RegistrationError(deviceData.id, deviceData.type);
-            }
-
-            callback(errorObj);
-        }
-    };
-}
-
-/**
- * Sends a Context Provider registration or unregistration request to the Context Broker. As registrations cannot be
- * removed, an unregistration consists of an update of the existent registration to make reduce its duration to
- * 1 second.
- *
- * The entity type, entity name and lazy attributes list are needed in order to send the registration:
- *
- * @param {Boolen} unregister       Indicates whether this registration is an unregistration or register.
- * @param {Object} deviceData       Object containing all the deviceData needed to send the registration.
- */
-function sendRegistrations(unregister, deviceData, callback) {
-    var options = {
-        url: 'http://' + config.getConfig().contextBroker.host + ':' +
-            config.getConfig().contextBroker.port + '/NGSI9/registerContext',
-        method: 'POST',
-        json: {
-            contextRegistrations: [
-                {
-                    entities: [
-                        {
-                            type: deviceData.type,
-                            isPattern: 'false',
-                            id: String(deviceData.name)
-                        }
-                    ],
-                    attributes: [],
-                    providingApplication: config.getConfig().providerUrl
-                }
-            ],
-            duration: (unregister) ? 'PT1S' : config.getConfig().deviceRegistrationDuration
-        },
-        headers: {
-            'fiware-service': deviceData.service,
-            'fiware-servicepath': deviceData.subservice
-        }
-    };
-
-    function formatAttributes(originalVector) {
-        var attributeList = [];
-
-        if (originalVector && originalVector.length) {
-            for (var i = 0; i < originalVector.length; i++) {
-                attributeList.push({
-                    name: originalVector[i].name,
-                    type: originalVector[i].type,
-                    isDomain: 'false'
-                });
-            }
-        }
-
-        return attributeList;
-    }
-
-    if (deviceData.registrationId) {
-        options.json.registrationId = deviceData.registrationId;
-    }
-
-    options.json.contextRegistrations[0].attributes = options.json.contextRegistrations[0].attributes.concat(
-        formatAttributes(deviceData.lazy),
-        formatAttributes(deviceData.commands)
-    );
-
-    if (options.json.contextRegistrations[0].attributes.length === 0) {
-        logger.debug(context, 'No Context Provider registrations found for unregister');
-        callback(null, deviceData);
-    } else {
-        logger.debug(context, 'Sending device registrations to Context Broker at [%s]', options.url);
-        logger.debug(context, 'Using the following request:\n\n%s\n\n', JSON.stringify(options, null, 4));
-
-        request(options, createRegistrationHandler(unregister, deviceData, callback));
-    }
-}
 
 /**
  * Process the response from a Register Context request for a device, extracting the 'registrationId' and creating the
@@ -351,7 +239,7 @@ function registerDevice(deviceObj, callback) {
         logger.debug(context, 'Registering device into NGSI Service:\n%s', JSON.stringify(deviceData, null, 4));
 
         async.waterfall([
-            apply(sendRegistrations, false, deviceData),
+            apply(registrationUtils.sendRegistrations, false, deviceData),
             apply(processContextRegistration, deviceData),
             apply(createInitialEntity, deviceData),
             config.getRegistry().store
@@ -393,7 +281,7 @@ function unregisterDevice(id, callback) {
         async.waterfall([
             async.apply(removeAllSubscriptions, device),
             processUnsubscribes,
-            async.apply(sendRegistrations, true, device),
+            async.apply(registrationUtils.sendRegistrations, true, device),
             processContextUnregister,
             async.apply(config.getRegistry().remove, id)
         ], callback);
@@ -472,7 +360,7 @@ function updateRegisterDevice(deviceObj, callback) {
         async.apply(extractDeviceDifference, deviceObj),
         createInitialEntity,
         async.apply(combineWithNewDevice, deviceObj),
-        async.apply(sendRegistrations, false),
+        async.apply(registrationUtils.sendRegistrations, false),
         async.apply(processContextRegistration, deviceObj),
         config.getRegistry().update
     ], callback);
